@@ -13,18 +13,16 @@ class GarbargeCollectionSuite:
     # custom because `memory_profiler` is a line-by-line profiler (also: https://github.com/pythonprofilers/memory_profiler/issues/402)
     def track_peakmem_garbage_collection(self, *_):
         def display_top(snapshot, key_type="lineno"):
-            snapshot = snapshot.filter_traces(
-                (
-                    tracemalloc.Filter(
-                        inclusive=False,
-                        filename_pattern="<frozen importlib._bootstrap>",
-                    ),
-                    tracemalloc.Filter(
-                        inclusive=False,
-                        filename_pattern="<unknown>",
-                    ),
-                )
-            )
+            snapshot = snapshot.filter_traces((
+                tracemalloc.Filter(
+                    inclusive=False,
+                    filename_pattern="<frozen importlib._bootstrap>",
+                ),
+                tracemalloc.Filter(
+                    inclusive=False,
+                    filename_pattern="<unknown>",
+                ),
+            ))
             top_stats = snapshot.statistics(key_type)
             total = sum(stat.size for stat in top_stats)
             return total
