@@ -1,14 +1,12 @@
 from __future__ import annotations
 
+from collections.abc import Callable
 from functools import WRAPPER_ASSIGNMENTS, wraps
 from itertools import pairwise
-from typing import TYPE_CHECKING, cast
-from warnings import warn
-
-import h5py
-from packaging.version import Version
+from typing import TYPE_CHECKING, Literal, cast
 
 from .._core.sparse_dataset import BaseCompressedSparseDataset
+from ..utils import warn
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Mapping
@@ -21,9 +19,6 @@ if TYPE_CHECKING:
 
     Storage = StorageType | BaseCompressedSparseDataset
 
-# For allowing h5py v3
-# https://github.com/scverse/anndata/issues/442
-H5PY_V3 = Version(h5py.__version__).major >= 3
 
 # -------------------------------------------------------------------------------
 # Type conversion
@@ -133,7 +128,7 @@ def read_attribute(*args, **kwargs):
     from .specs import read_elem
 
     msg = "This internal function has been deprecated, please use read_elem instead"
-    warn(msg, FutureWarning, stacklevel=2)
+    warn(msg, FutureWarning)
     return read_elem(*args, **kwargs)
 
 
@@ -141,7 +136,7 @@ def write_attribute(*args, **kwargs):
     from .specs import write_elem
 
     msg = "This internal function has been deprecated, please use write_elem instead"
-    warn(msg, FutureWarning, stacklevel=2)
+    warn(msg, FutureWarning)
     return write_elem(*args, **kwargs)
 
 
