@@ -2,12 +2,10 @@ from collections.abc import Callable as Callable
 from collections.abc import Generator, Iterable
 from contextlib import contextmanager
 from dataclasses import dataclass
-from typing import Literal, TypeVar
-
-_T = TypeVar("_T")
+from typing import Literal, Self
 
 @dataclass
-class SettingsManager:
+class SettingsManager[T]:
     __doc_tmpl__: str = ...
     def describe(
         self,
@@ -23,11 +21,11 @@ class SettingsManager:
         self,
         option: str,
         *,
-        default_value: _T,
+        default_value: T,
         description: str,
-        validate: Callable[[_T], None],
+        validate: Callable[[T, Self], None],
         option_type: object | None = None,
-        get_from_env: Callable[[str, _T], _T] = ...,
+        get_from_env: Callable[[str, T], T] = ...,
     ) -> None: ...
     def __setattr__(self, option: str, val: object) -> None: ...
     def __getattr__(self, option: str) -> object: ...
@@ -45,5 +43,7 @@ class _AnnDataSettingsManager(SettingsManager):
     zarr_write_format: Literal[2, 3] = 2
     use_sparse_array_on_read: bool = False
     min_rows_for_chunked_h5_copy: int = 1000
+    disallow_forward_slash_in_h5ad: bool = False
+    auto_shard_zarr_v3: bool = False
 
 settings: _AnnDataSettingsManager
