@@ -3,19 +3,22 @@ from __future__ import annotations
 from codecs import decode
 from collections.abc import Mapping, Sequence
 from functools import cache, partial, singledispatch
+from importlib.metadata import version
 from importlib.util import find_spec
 from types import EllipsisType
-from typing import TYPE_CHECKING, TypeVar
-from warnings import warn
+from typing import TYPE_CHECKING
 
 import h5py
 import numpy as np
 import pandas as pd
-import scipy
+import scipy.sparse
+from legacy_api_wrap import legacy_api  # noqa: TID251
 from numpy.typing import NDArray
 from packaging.version import Version
 from zarr import Array as ZarrArray  # noqa: F401
 from zarr import Group as ZarrGroup
+
+from .._warnings import warn
 
 if TYPE_CHECKING:
     from typing import Any
@@ -75,15 +78,14 @@ H5File = h5py.File
 #############################
 @cache
 def is_zarr_v2() -> bool:
-    import zarr
     from packaging.version import Version
 
-    return Version(zarr.__version__) < Version("3.0.0")
+    return Version(version("zarr")) < Version("3.0.0")
 
 
 if is_zarr_v2():
     msg = "anndata will no longer support zarr v2 in the near future. Please prepare to upgrade to zarr>=3."
-    warn(msg, DeprecationWarning, stacklevel=2)
+    warn(msg, DeprecationWarning)
 
 
 if find_spec("awkward") or TYPE_CHECKING:
@@ -198,14 +200,7 @@ else:
 
 CupyCSMatrix = CupyCSCMatrix | CupyCSRMatrix
 
-if find_spec("legacy_api_wrap") or TYPE_CHECKING:
-    from legacy_api_wrap import legacy_api  # noqa: TID251
-
-    old_positionals = partial(legacy_api, category=FutureWarning)
-else:
-
-    def old_positionals(*old_positionals):
-        return lambda func: func
+old_positionals = partial(legacy_api, category=FutureWarning)
 
 
 #############################
@@ -215,7 +210,7 @@ else:
 
 NULLABLE_NUMPY_STRING_TYPE = (
     np.dtype("O")
-    if Version(np.__version__) < Version("2")
+    if Version(version("numpy")) < Version("2")
     else np.dtypes.StringDType(na_object=pd.NA)
 )
 
@@ -327,12 +322,9 @@ def _to_fixed_length_strings(value: np.ndarray) -> np.ndarray:
     return value.astype(new_dtype)
 
 
-Group_T = TypeVar("Group_T", bound=ZarrGroup | h5py.Group)
-
-
 # TODO: This is a workaround for https://github.com/scverse/anndata/issues/874
 # See https://github.com/h5py/h5py/pull/2311#issuecomment-1734102238 for why this is done this way.
-def _require_group_write_dataframe(
+def _require_group_write_dataframe[Group_T: ZarrGroup | h5py.Group](
     f: Group_T, name: str, df: pd.DataFrame, *args, **kwargs
 ) -> Group_T:
     if len(df.columns) > 5_000 and isinstance(f, H5Group):
@@ -373,10 +365,8 @@ def _clean_uns(adata: AnnData):  # noqa: F821
         del adata.uns[cats_name]
 
 
-def _move_adj_mtx(d):
-    """
-    Read-time fix for moving adjacency matrices from uns to obsp
-    """
+def _move_adj_mtx(d) -> None:
+    """Read-time fix for moving adjacency matrices from uns to obsp."""
     n = d.get("uns", {}).get("neighbors", {})
     obsp = d.setdefault("obsp", {})
 
@@ -390,8 +380,7 @@ def _move_adj_mtx(d):
                 f"Moving element from .uns['neighbors'][{k!r}] to .obsp[{k!r}].\n\n"
                 "This is where adjacency matrices should go now."
             )
-            # 5: caller -> 4: legacy_api_wrap -> 3: `AnnData.__init__` -> 2: `_init_as_actual` → 1: here
-            warn(msg, FutureWarning, stacklevel=5)
+            warn(msg, FutureWarning)
             obsp[k] = n.pop(k)
 
 
@@ -430,11 +419,3 @@ def _safe_transpose(x):
         return _transpose_by_block(x)
     else:
         return x.T
-
-
-def _map_cat_to_str(cat: pd.Categorical) -> pd.Categorical:
-    if Version(pd.__version__) >= Version("2.1"):
-        # Argument added in pandas 2.1
-        return cat.map(str, na_action="ignore")
-    else:
-        return cat.map(str)
