@@ -1,9 +1,9 @@
 from __future__ import annotations
 
 from functools import partial
+from importlib.metadata import version
 from typing import TYPE_CHECKING
 
-import dask
 import joblib
 import pytest
 from dask.base import normalize_token, tokenize
@@ -11,7 +11,7 @@ from packaging.version import Version
 
 from anndata.compat import is_zarr_v2
 
-if Version(dask.__version__) < Version("2024.8.0"):
+if Version(version("dask")) < Version("2024.8.0"):
     from dask.base import normalize_seq
 else:
     from dask.tokenize import normalize_seq
@@ -118,11 +118,14 @@ def local_cluster_addr(
     # Adapted from https://pytest-xdist.readthedocs.io/en/latest/how-to.html#making-session-scoped-fixtures-execute-only-once
     import dask.distributed as dd
 
-    def make_cluster() -> dd.LocalCluster:
-        return dd.LocalCluster(n_workers=1, threads_per_worker=1)
+    def make_cluster(worker_id: str) -> dd.LocalCluster:
+        # If we're not using multiple pytest-xdist workers, let the cluster have multiple workers.
+        return dd.LocalCluster(
+            n_workers=1 if worker_id != "master" else 2, threads_per_worker=1
+        )
 
     if worker_id == "master":
-        with make_cluster() as cluster:
+        with make_cluster(worker_id) as cluster:
             yield cluster.scheduler_address
             return
 
@@ -138,7 +141,7 @@ def local_cluster_addr(
         yield address
         return
 
-    with make_cluster() as cluster:
+    with make_cluster(worker_id) as cluster:
         fn.write_text(cluster.scheduler_address)
         lock.release()
         yield cluster.scheduler_address
