@@ -508,17 +508,17 @@ def test_append_col():
     # adata.obs[['new2', 'new3']] = [['A', 'B'], ['c', 'd']]
 
     with pytest.raises(
-        ValueError, match="Length of values.*does not match length of index"
+        ValueError, match=r"Length of values.*does not match length of index"
     ):
         adata.obs["new4"] = ["far", "too", "long"]
 
 
 def test_delete_col():
     adata = AnnData(np.array([[1, 2, 3], [4, 5, 6]]), dict(o1=[1, 2], o2=[3, 4]))
-    assert adata.obs_keys() == ["o1", "o2"]
+    assert adata.obs.columns.tolist() == ["o1", "o2"]
 
     del adata.obs["o1"]
-    assert adata.obs_keys() == ["o2"]
+    assert adata.obs.columns.tolist() == ["o2"]
     assert adata.obs["o2"].tolist() == [3, 4]
 
 
@@ -536,7 +536,7 @@ def test_multicol():
     adata = AnnData(np.array([[1, 2, 3], [4, 5, 6]]))
     # 'c' keeps the columns as should be
     adata.obsm["c"] = np.array([[0.0, 1.0], [2, 3]])
-    assert adata.obsm_keys() == ["c"]
+    assert adata.obsm.keys() == {"c"}
     assert adata.obsm["c"].tolist() == [[0.0, 1.0], [2, 3]]
 
 
