@@ -43,13 +43,13 @@ def test_read_partial_X(tmp_path, typ, diskfmt):
 
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
 def test_read_partial_adata(tmp_path, diskfmt):
+    import scanpy as sc
+
     with warnings.catch_warnings():
         warnings.filterwarnings(
-            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
+            "ignore", r"Moving element.*uns.*to.*obsp", FutureWarning
         )
-        import scanpy as sc
-
-    adata = sc.datasets.pbmc68k_reduced()
+        adata = sc.datasets.pbmc68k_reduced()
 
     path = Path(tmp_path) / ("test_rp." + diskfmt)
 
